Introduces concept for custom static pages.
Static pages can now simply be placed in application/modules/home/views/scripts/*.phtml. Nothing has to be done in the
controller, the page is automatically included in the menu. Take a look at application/modules/home/language/home.tmx
for an example of how to name the translation string for the menu entry to be displayed correctly. The basic idea, which
we might want to incorporate for all translation strings, is to prefix the strings with module-, controller-, and
actionname (e.g. 'home_index_about_sometext'). For the menu entries 'home_modulename', 'home_index_controllername' and
'home_index_about_actionname' should be properly set.

git-svn-id: https://svn.zib.de/opus4dev/application/trunk@1470 0a49f362-4213-0410-86e3-e96cb6c4fbae