Interim commit: Refactors collection controller.
One controller is now responsible for both, collections and roles. Collection(Role)s are now navigated per level, a
breadcrumb trail is displayed to clarify current position.

FIXME: Performance issues when saving collections. This would be a reason to implement advanced modification tracking
for multi-value fields (i.e. track not only that the field has changed, but also which items within the field need to
be updated.)

git-svn-id: https://svn.zib.de/opus4dev/application/trunk@2630 0a49f362-4213-0410-86e3-e96cb6c4fbae