Added support for *firstname*, *lastname*, and *email* for accounts. Added new option for clearing module:
{code}
clearing.addCurrentUserAsReferee = 1
{code}

The option can be used to disable the input form for referee information in the review module. I have turned
off the option in the *application.ini* for now, since it requires that the current user account has a firstname
and a lastname attribute. Otherwise the clearance module won't clear the documents.

We still need to decide how to deal with missing information in the current user account.

git-svn-id: https://svn.zib.de/opus4dev/server/trunk@6793 0a49f362-4213-0410-86e3-e96cb6c4fbae